docs , added exmaples, renamed grid_range module
