+++ pygsheets/__init__.py
@@ -16,7 +16,7 @@ from pygsheets.spreadsheet import Spreadsheet
 from pygsheets.worksheet import Worksheet
 from pygsheets.cell import Cell
 from pygsheets.datarange import DataRange
-from pygsheets.grid_range import GridRange, Address
+from pygsheets.address import GridRange, Address
 from pygsheets.chart import Chart
 from pygsheets.utils import format_addr
 from pygsheets.custom_types import (FormatType, WorkSheetProperty, DateTimeRenderOption,
